chore: port new test modules to separate file
