+++ src/bn254/fq12.rs
@@ -858,11 +858,8 @@ impl Fq12 {
 
 #[cfg(test)]
 mod test {
-    use crate::bn254::fp254impl::Fp254Impl;
-    use crate::bn254::fq::Fq;
     use crate::bn254::fq12::Fq12;
-    use crate::bn254::fq6::Fq6;
-    use crate::bn254::utils::{fq12_push, fq12_push_not_montgomery, fq2_push, fq2_push_not_montgomery, hash_fp12, hash_fp12_192, hash_fp12_with_hints, hash_fp6};
+    use crate::bn254::utils::{fq12_push, fq12_push_not_montgomery, fq2_push, fq2_push_not_montgomery};
     use crate::{execute_script_without_stack_limit, treepp::*};
     use ark_ff::AdditiveGroup;
     use ark_ff::{CyclotomicMultSubgroup, Field};
@@ -936,202 +933,6 @@ mod test {
             };
             run(script);
         }
-    }
-
-
-    #[test]
-    fn test_bn254_fq12_hinted_mul_split0() {
-        let mut prng: ChaCha20Rng = ChaCha20Rng::seed_from_u64(0);
-
-        let mut max_stack = 0;
-
-        for _ in 0..1 {
-            let a = ark_bn254::Fq12::rand(&mut prng);
-            let b = ark_bn254::Fq12::rand(&mut prng);
-            let c = a.mul(&b);
-
-            let (hinted_mul, hints) = Fq12::hinted_mul_first(12, a, 0, b);
-
-            let script = script! {
-                for hint in hints { 
-                    { hint.push() }
-                }
-                // Hash_b
-                {u32::from_le_bytes([17, 50, 164, 0])}
-                {u32::from_le_bytes([235, 77, 217, 15])}
-                {u32::from_le_bytes([1, 4, 86, 10])}
-                {u32::from_le_bytes([23, 225, 110, 26])}
-                {u32::from_le_bytes([71, 105, 236, 11])}
-                {u32::from_le_bytes([75, 29, 151, 8])}
-                {u32::from_le_bytes([130, 190, 188, 3])}
-                {u32::from_le_bytes([246, 67, 44, 19])}
-                {u32::from_le_bytes([105, 194, 20, 27])}
-
-                // Hash_a
-                {u32::from_le_bytes([131, 116, 114, 0])}
-                {u32::from_le_bytes([245, 129, 139, 3])}
-                {u32::from_le_bytes([132, 171, 199, 7])}
-                {u32::from_le_bytes([97, 185, 93, 16])}
-                {u32::from_le_bytes([161, 222, 150, 25])}
-                {u32::from_le_bytes([44, 144, 71, 23])}
-                {u32::from_le_bytes([139, 185, 38, 22])}
-                {u32::from_le_bytes([233, 138, 103, 22])}
-                {u32::from_le_bytes([9, 213, 155, 19])}
-                
-                // Hash_c
-                {u32::from_le_bytes([82, 143, 25,0])}
-                {u32::from_le_bytes([106, 69, 151, 13])}
-                {u32::from_le_bytes([154, 120, 131, 27])}
-                {u32::from_le_bytes([140, 55, 239, 25])}
-                {u32::from_le_bytes([92, 201, 47, 28])}
-                {u32::from_le_bytes([44, 174, 74, 16])}
-                {u32::from_le_bytes([57, 190, 31, 19])}
-                {u32::from_le_bytes([31, 231, 126, 1])}
-                {u32::from_le_bytes([158, 11, 210, 2])}
-
-                { fq12_push_not_montgomery(a) }
-                { fq12_push_not_montgomery(b) } // fp12_one
-                { hinted_mul.clone() }
-                { Fq6::toaltstack() }
-                
-                { hash_fp12()}
-                //bring Hashb to top
-                for i in 0..9 {
-                    OP_DEPTH OP_1SUB OP_ROLL
-                }
-                { Fq::equalverify(0, 1)}
-
-                // hash_a
-                { hash_fp12_192() }
-                for i in 0..9 {
-                    OP_DEPTH OP_1SUB OP_ROLL
-                }
-                { Fq::equalverify(0, 1)}
-
-                {Fq::fromaltstack()} // Fq_claimed from altstack
-                {Fq::equalverify(0, 1)} // SHOULD BE UNEQUAL VERIFY
-                OP_TRUE
-            };
-
-            println!("script len {}", script.len());
-            let exec_result = execute_script(script);
-            assert!(exec_result.success);
-            for i in 0..exec_result.final_stack.len() {
-                println!("{i:3} {:?}", exec_result.final_stack.get(i));
-            }
-            max_stack = max_stack.max(exec_result.stats.max_nb_stack_items);
-            println!("Fq12::mul {} stack", max_stack);
-            
-        }
-
-    }
-
-
-
-    #[test]
-    fn test_bn254_fq12_hinted_mul_split1() {
-        let mut prng: ChaCha20Rng = ChaCha20Rng::seed_from_u64(0);
-
-        let mut max_stack = 0;
-
-        for _ in 0..1 {
-            let a = ark_bn254::Fq12::rand(&mut prng);
-            let b = ark_bn254::Fq12::rand(&mut prng);
-            let c = a.mul(&b);
-
-            let (hinted_mul, hints) = Fq12::hinted_mul_second(12, a, 0, b);
-
-            let script = script! {
-                for hint in hints { 
-                    { hint.push() }
-                }
-                // Hash_b
-                {u32::from_le_bytes([17, 50, 164, 0])}
-                {u32::from_le_bytes([235, 77, 217, 15])}
-                {u32::from_le_bytes([1, 4, 86, 10])}
-                {u32::from_le_bytes([23, 225, 110, 26])}
-                {u32::from_le_bytes([71, 105, 236, 11])}
-                {u32::from_le_bytes([75, 29, 151, 8])}
-                {u32::from_le_bytes([130, 190, 188, 3])}
-                {u32::from_le_bytes([246, 67, 44, 19])}
-                {u32::from_le_bytes([105, 194, 20, 27])}
-
-                // Hash_c0
-                {u32::from_le_bytes([17, 50, 164, 0])}
-                {u32::from_le_bytes([235, 77, 217, 15])}
-                {u32::from_le_bytes([1, 4, 86, 10])}
-                {u32::from_le_bytes([23, 225, 110, 26])}
-                {u32::from_le_bytes([71, 105, 236, 11])}
-                {u32::from_le_bytes([75, 29, 151, 8])}
-                {u32::from_le_bytes([130, 190, 188, 3])}
-                {u32::from_le_bytes([246, 67, 44, 19])}
-                {u32::from_le_bytes([105, 194, 20, 27])}
-
-                // Hash_a
-                {u32::from_le_bytes([131, 116, 114, 0])}
-                {u32::from_le_bytes([245, 129, 139, 3])}
-                {u32::from_le_bytes([132, 171, 199, 7])}
-                {u32::from_le_bytes([97, 185, 93, 16])}
-                {u32::from_le_bytes([161, 222, 150, 25])}
-                {u32::from_le_bytes([44, 144, 71, 23])}
-                {u32::from_le_bytes([139, 185, 38, 22])}
-                {u32::from_le_bytes([233, 138, 103, 22])}
-                {u32::from_le_bytes([9, 213, 155, 19])}
-                
-                // Hash_c
-                {u32::from_le_bytes([66, 234, 4, 0])}
-                {u32::from_le_bytes([156, 104, 70, 7])}
-                {u32::from_le_bytes([5, 60, 102, 10])}
-                {u32::from_le_bytes([171, 108, 80, 11])}
-                {u32::from_le_bytes([30, 94, 254, 19])}
-                {u32::from_le_bytes([34, 232, 58, 11])}
-                {u32::from_le_bytes([191, 101, 160, 16])}
-                {u32::from_le_bytes([53, 186, 189, 25])}
-                {u32::from_le_bytes([83, 33, 154, 8])}
-
-                { fq12_push_not_montgomery(a) }
-                { fq12_push_not_montgomery(b) } // fp12_one
-                { hinted_mul.clone() }
-                { Fq6::toaltstack() }
-                
-                { hash_fp12()}
-                // bring Hashb to top
-                for i in 0..9 {
-                    OP_DEPTH OP_1SUB OP_ROLL
-                }
-                { Fq::equalverify(0, 1)}
-
-                // bring Hash_c0 to top
-                for i in 0..9 {
-                    OP_DEPTH OP_1SUB OP_ROLL
-                }
-                { Fq6::fromaltstack() }
-                { hash_fp12_with_hints() }
-                { Fq::toaltstack() } // Fq_claimed to altstack
-
-                // hash_a
-                { hash_fp12_192() }
-                for i in 0..9 {
-                    OP_DEPTH OP_1SUB OP_ROLL
-                }
-                { Fq::equalverify(0, 1)}
-
-                {Fq::fromaltstack()} // Fq_claimed from altstack
-                {Fq::equalverify(0, 1)} // SHOULD BE UNEQUAL VERIFY
-                OP_TRUE
-            };
-
-            println!("script len {}", script.len());
-            let exec_result = execute_script(script);
-            assert!(exec_result.success);
-            for i in 0..exec_result.final_stack.len() {
-                println!("{i:3} {:?}", exec_result.final_stack.get(i));
-            }
-            max_stack = max_stack.max(exec_result.stats.max_nb_stack_items);
-            println!("Fq12::mul {} stack", max_stack);
-            
-        }
-
     }
 
 
@@ -1269,76 +1070,6 @@ mod test {
             run(script);
         }
     }
-
-
-    #[test]
-    fn test_bn254_fq12_hinted_square() {
-        let mut prng = ChaCha20Rng::seed_from_u64(0);
-
-        let mut max_stack = 0u32;
-
-
-        for _ in 0..1 {
-            let a = ark_bn254::Fq12::rand(&mut prng);
-            let b = a.square();
-
-            let (hinted_square, hints) = Fq12::hinted_square(a);
-
-            println!("hints len {:?}", hints.len());
-            let script = script! {
-           
-                for hint in hints { 
-                    { hint.push() }
-                }
-                {u32::from_le_bytes([234, 190, 118, 0])}
-                {u32::from_le_bytes([135, 162, 241, 5])}
-                {u32::from_le_bytes([186, 52, 155, 0])}
-                {u32::from_le_bytes([162, 11, 254, 31])}
-                {u32::from_le_bytes([130, 167, 61, 5])}
-                {u32::from_le_bytes([178, 14, 103, 23])}
-                {u32::from_le_bytes([195, 223, 134, 7])}
-                {u32::from_le_bytes([68, 150, 213, 11])}
-                {u32::from_le_bytes([83, 1, 19, 0])}
-
-                {u32::from_le_bytes([234, 190, 118, 0])}
-                {u32::from_le_bytes([135, 162, 241, 5])}
-                {u32::from_le_bytes([186, 52, 155, 0])}
-                {u32::from_le_bytes([162, 11, 254, 31])}
-                {u32::from_le_bytes([130, 167, 61, 5])}
-                {u32::from_le_bytes([178, 14, 103, 23])}
-                {u32::from_le_bytes([195, 223, 134, 7])}
-                {u32::from_le_bytes([68, 150, 213, 11])}
-                {u32::from_le_bytes([83, 1, 19, 0])}
-
-                { fq12_push_not_montgomery(a) }
-                { fq12_push_not_montgomery(a) }
-                { hinted_square.clone() }
-                { hash_fp12() }
-
-                { Fq::drop() }
-
-                { hash_fp12() }
-
-                {Fq::drop()}
-                {Fq::drop()}
-                {Fq::drop()}
-                OP_TRUE
-
-            };
-            println!("len {:?}", script.len());
-            let exec_result = execute_script(script);
-
-            //assert!(exec_result.success);
-            for i in 0..exec_result.final_stack.len() {
-                println!("{i:3} {:?}", exec_result.final_stack.get(i));
-            }
-            println!("stack len {:?} final len {:?}", exec_result.stats.max_nb_stack_items, exec_result.final_stack.len());
-            // max_stack = max_stack.max(exec_result.stats.max_nb_stack_items);
-            // println!("Fq12::hinted_square: {} @ {} stack final stack {}", hinted_square.len(), max_stack,exec_result.final_stack.len());
-            println!("len12 {} len12h {} len6 {}", hash_fp12().len(), hash_fp12_with_hints().len(), hash_fp6().len());
-        }
-    }
-
 
     #[test]
     fn test_bn254_fq12_mul_by_034() {
