+++ src/bn254/utils.rs
@@ -1365,210 +1365,6 @@ pub fn double_line() -> Script {
     }
 }
 
-use std::fs::File;
-use std::io::{self, Read};
-pub fn read_script_from_file(file_path: &str) -> Script {
-    fn read_file_to_bytes(file_path: &str) -> io::Result<Vec<u8>> {
-        let mut file = File::open(file_path)?;
-        let mut all_script_bytes = Vec::new();
-        file.read_to_end(&mut all_script_bytes)?;
-        Ok(all_script_bytes)
-    }
-    //let file_path = "blake3_bin/blake3_192b_252k.bin"; // Replace with your file path
-    let all_script_bytes = read_file_to_bytes(file_path).unwrap();
-    let scb = ScriptBuf::from_bytes(all_script_bytes);
-    let sc = script!();
-    let sc = sc.push_script(scb);
-    sc
-}
-
-
-// [a0, a1, a2, a3, a4, a5]
-// [H(a0,a1), H(a2,a3,a4,a5)]
-// [Hb0, Hb1]
-// [Hb1, Hb0]
-// Hash(Hb1, Hb0)
-// Hb
-pub fn hash_fp6() -> Script {
-
-    let hash_64b_75k = read_script_from_file("blake3_bin/blake3_64b_75k.bin");
-    let hash_128b_168k = read_script_from_file("blake3_bin/blake3_128b_168k.bin");
-
-    script!{
-        for _ in 0..=4 {
-            {Fq::toaltstack()}
-        }
-        { unpack_u32_to_u8() }
-        { Fq::fromaltstack() }
-        { unpack_u32_to_u8() }
-        {hash_64b_75k.clone()}
-        { pack_u8_to_u32() }
-
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { hash_128b_168k.clone() }
-
-            for _ in 0..9 {
-                {64 + 8} OP_ROLL
-            }
-            {unpack_u32_to_u8()}
-        { hash_64b_75k }
-        {pack_u8_to_u32()}
-    } 
-}
-
-pub fn hash_fp12() -> Script {
-
-    let hash_64b_75k = read_script_from_file("blake3_bin/blake3_64b_75k.bin");
-    let hash_128b_168k = read_script_from_file("blake3_bin/blake3_128b_168k.bin");
-
-    script!{
-        for _ in 0..=10 {
-            {Fq::toaltstack()}
-        }
-
-        // first part
-        { unpack_u32_to_u8() }
-        { Fq::fromaltstack() }
-        { unpack_u32_to_u8() }
-        {hash_64b_75k.clone()}
-        { pack_u8_to_u32() }
-
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { hash_128b_168k.clone() }
-
-
-        for _ in 0..9 {
-            {64 + 8} OP_ROLL
-        }
-        {unpack_u32_to_u8()}
-        {hash_64b_75k.clone()}
-        {pack_u8_to_u32()}
-
-        // second part
-
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        {hash_64b_75k.clone()}
-        { pack_u8_to_u32() }
-        
-
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { hash_128b_168k.clone() }
-
-        for _ in 0..9 {
-            {64 + 8} OP_ROLL
-        }
-        {unpack_u32_to_u8()}
-        {hash_64b_75k.clone()}
-
-        // wrap up
-        for _ in 0..9 {
-            {64 + 8} OP_ROLL
-        }
-        {unpack_u32_to_u8()}
-        {hash_64b_75k.clone()}
-        {pack_u8_to_u32()}
-
-    } 
-}
-
-pub fn hash_fp12_192() -> Script {
-    let hash_64b_75k = read_script_from_file("blake3_bin/blake3_64b_75k.bin");
-    let hash_192b_252k = read_script_from_file("blake3_bin/blake3_192b_252k.bin");
-    script! {
-        for _ in 0..=10 {
-            {Fq::toaltstack()}
-        }
-        {unpack_u32_to_u8() }
-        for _ in 0..5 {
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-        }
-        {hash_192b_252k.clone()}
-        {pack_u8_to_u32()}
-
-        for _ in 0..6 {
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-        }
-        {hash_192b_252k}
-        for _ in 0..9 {
-            {64+8} OP_ROLL
-        }
-        { unpack_u32_to_u8() }
-        {hash_64b_75k}
-        {pack_u8_to_u32()}
-    }
-}
-
-// 6Fp_hash
-// fp6
-pub fn hash_fp12_with_hints() -> Script {
-
-    let hash_64b_75k = read_script_from_file("blake3_bin/blake3_64b_75k.bin");
-    let hash_128b_168k = read_script_from_file("blake3_bin/blake3_128b_168k.bin");
-
-    script!{
-        for _ in 0..=4 {
-            {Fq::toaltstack()}
-        }
-
-        { unpack_u32_to_u8() }
-        { Fq::fromaltstack() }
-        { unpack_u32_to_u8() }
-        {hash_64b_75k.clone()}
-        { pack_u8_to_u32() }
-
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { Fq::fromaltstack() }
-        {unpack_u32_to_u8()}
-        { hash_128b_168k.clone() }
-
-
-        for _ in 0..9 {
-            {64 + 8} OP_ROLL
-        }
-        {unpack_u32_to_u8()}
-        {hash_64b_75k.clone()}
-
-        // wrap up
-        for _ in 0..9 {
-            {64 + 8} OP_ROLL
-        }
-        {unpack_u32_to_u8()}
-        {hash_64b_75k.clone()}
-        {pack_u8_to_u32()}
-
-    } 
-}
 
 
 #[cfg(test)]
@@ -1830,86 +1626,6 @@ mod test {
     }
 
     #[test]
-    fn test_hinited_sparse_dense_mul() {
-        let mut prng = ChaCha20Rng::seed_from_u64(0);
-        let t = ark_bn254::G2Affine::rand(&mut prng);
-        let _q = ark_bn254::G2Affine::rand(&mut prng);
-        let p = ark_bn254::G1Affine::rand(&mut prng);
-        let f = ark_bn254::Fq12::rand(&mut prng);
-        
-        let two_inv = ark_bn254::Fq::one().double().inverse().unwrap();
-        let three_div_two = (ark_bn254::Fq::one().double() + ark_bn254::Fq::one()) * two_inv;
-        let mut alpha_tangent = t.x.square();
-        alpha_tangent /= t.y;
-        alpha_tangent.mul_assign_by_fp(&three_div_two);
-        // -bias
-        let bias_minus_tangent = alpha_tangent * t.x - t.y;
-
-        let mut f1 = f;
-        let mut c1new = alpha_tangent;
-        c1new.mul_assign_by_fp(&(-p.x / p.y));
-
-        let mut c2new = bias_minus_tangent;
-        c2new.mul_assign_by_fp(&(p.y.inverse().unwrap()));
-
-        f1.mul_by_034(&ark_bn254::Fq2::ONE, &c1new, &c2new);
-        let (hinted_script, hints) = Fq12::hinted_mul_by_34(f, c1new, c2new);
-
-        let hash_128b_168k = read_script_from_file("blake3_bin/blake3_128b_168k.bin");
-
-
-        let script = script! {
-            for hint in hints { 
-                { hint.push() }
-            }
-
-            { fq12_push_not_montgomery(f) }
-            { fq2_push_not_montgomery(c1new) }
-            { fq2_push_not_montgomery(c2new) }
-            { fq12_push_not_montgomery(f) }
-            { fq2_push_not_montgomery(c1new) }
-            { fq2_push_not_montgomery(c2new) }
-            { hinted_script.clone() }
-
-            for _ in 0..12 { // save claim
-                {Fq::toaltstack()}
-            }
-
-            // hash c1new and c2new
-            {Fq::roll(1)}
-            {Fq::roll(2)}
-            {Fq::roll(3)}
-            { Fq::toaltstack() }
-            { Fq::toaltstack() }
-            { Fq::toaltstack() }
-            {unpack_u32_to_u8()} // 0
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            {hash_128b_168k.clone()}
-            {pack_u8_to_u32()}
-
-            {hash_fp12() }
-
-            for _ in 0..12 { // save claim
-                {Fq::fromaltstack()}
-            }
-            { hash_fp12_192() }
-
-            OP_TRUE
-        };
-        let len = script.len();
-        let res = execute_script(script);
-        assert!(res.success);
-        let mut max_stack = 0;
-        max_stack = max_stack.max(res.stats.max_nb_stack_items);
-        println!("Fq6::window_mul: {} @ {} stack", len, max_stack);
-    }
-
-    #[test]
     fn test_hinted_sparse_mul() {
         // take fixed Qs as input vkY and vk_del
         let mut prng = ChaCha20Rng::seed_from_u64(0);
@@ -2026,277 +1742,6 @@ mod test {
         let (scd, (nt2, nt3)) = sparse_mul_for_doubling(t2, t3);
 
         let (sca, (nt2, nt3)) = sparse_mul_for_addition(nt2, nt3, q_vk_g, q_vk_del);
-    }
-
-    #[test]
-    fn test_hinted_affine_double_add_eval() {
-        // slope: alpha = 3 * x^2 / 2 * y
-        // intercept: bias = y - alpha * x
-        // x' = alpha^2 - 2 * x
-        // y' = -bias - alpha * x'
-        let mut prng = ChaCha20Rng::seed_from_u64(0);
-        let t = ark_bn254::G2Affine::rand(&mut prng);
-        let two_inv = ark_bn254::Fq::one().double().inverse().unwrap();
-        let three_div_two = (ark_bn254::Fq::one().double() + ark_bn254::Fq::one()) * two_inv;
-        let mut alpha_tangent = t.x.square();
-        alpha_tangent /= t.y;
-        alpha_tangent.mul_assign_by_fp(&three_div_two);
-        // -bias
-        let bias_minus_tangent = alpha_tangent * t.x - t.y;
-
-        let x = alpha_tangent.square() - t.x.double();
-        let y = bias_minus_tangent - alpha_tangent * x;
-        let (hinted_double_line, hints_double_line) = new_hinted_affine_double_line(t.x, alpha_tangent, bias_minus_tangent);
-        let (hinted_check_tangent, hints_check_tangent) = new_hinted_check_line_through_point(t.x, alpha_tangent, bias_minus_tangent);
-
-        let tt = G2Affine::new(x, y);
-        let q = ark_bn254::G2Affine::rand(&mut prng);
-        let alpha_chord = (tt.y - q.y) / (tt.x - q.x);
-        // -bias
-        let bias_minus_chord = alpha_chord * tt.x - tt.y;
-        assert_eq!(alpha_chord * tt.x - tt.y, bias_minus_chord);
-
-        let x = alpha_chord.square() - tt.x - q.x;
-        let y = bias_minus_chord - alpha_chord * x;
-        let p = ark_bn254::g1::G1Affine::rand(&mut prng);
-        let p_dash_x = -p.x/p.y;
-        let p_dash_y = p.y.inverse().unwrap();
-
-        let (hinted_check_chord_t, hints_check_chord_t) = new_hinted_check_line_through_point( tt.x, alpha_chord, bias_minus_chord);
-        let (hinted_check_chord_q, hints_check_chord_q) = new_hinted_check_line_through_point( q.x, alpha_chord, bias_minus_chord);
-        let (hinted_add_line, hints_add_line) = new_hinted_affine_add_line(tt.x, q.x, alpha_chord, bias_minus_chord);
-
-        // affine mode as well
-        let mut c1new = alpha_tangent;
-        c1new.mul_assign_by_fp(&(-p.x / p.y));
-
-        let mut c2new = bias_minus_tangent;
-        c2new.mul_assign_by_fp(&(p.y.inverse().unwrap()));
-
-        let mut c1new_2 = alpha_chord;
-        c1new_2.mul_assign_by_fp(&(-p.x / p.y));
-
-        let mut c2new_2 = bias_minus_chord;
-        c2new_2.mul_assign_by_fp(&(p.y.inverse().unwrap()));
-
-        let (hinted_ell_tangent, hints_ell_tangent) = new_hinted_ell_by_constant_affine(p_dash_x, p_dash_y, alpha_tangent, bias_minus_tangent);
-        let (hinted_ell_chord, hints_ell_chord) = new_hinted_ell_by_constant_affine(p_dash_x, p_dash_y, alpha_chord, bias_minus_chord);
-
-        let hash_64b_75k = read_script_from_file("blake3_bin/blake3_64b_75k.bin");
-        let hash_128b_168k = read_script_from_file("blake3_bin/blake3_128b_168k.bin");
-
-        let bcsize = 6+3;
-        let script = script! {
-            // hints
-            for hint in hints_check_tangent { 
-                { hint.push() }
-            }
-            for hint in hints_ell_tangent { 
-                { hint.push() }
-            }
-            for hint in hints_double_line { 
-                { hint.push() }
-            }
-            for hint in hints_check_chord_q { 
-                { hint.push() }
-            }
-            for hint in hints_check_chord_t { 
-                { hint.push() }
-            }
-            for hint in hints_ell_chord { 
-                { hint.push() }
-            }
-            for hint in hints_add_line { 
-                { hint.push() }
-            }
-
-            // aux
-            { fq2_push_not_montgomery(alpha_chord)}
-            { fq2_push_not_montgomery(bias_minus_chord)}
-            { fq2_push_not_montgomery(alpha_tangent)}
-            { fq2_push_not_montgomery(bias_minus_tangent)}
-            { fq2_push_not_montgomery(t.x) }
-            { fq2_push_not_montgomery(t.y) }
-
-            // bit commits
-            { fq_push_not_montgomery(p_dash_x) }
-            { fq_push_not_montgomery(p_dash_y) }
-            { fq2_push_not_montgomery(q.x) }
-            { fq2_push_not_montgomery(q.y) }
-            { Fq::push_zero() } // hash
-            { Fq::push_zero() } // hash
-            { Fq::push_zero() } // hash
-            
-
-            { Fq2::copy(bcsize+6)} // alpha
-            { Fq2::copy(bcsize+6)} // bias
-            { Fq2::copy(bcsize+6)} // t.x
-            { Fq2::copy(bcsize+6)} // t.y
-            { hinted_check_tangent }
-
-            { Fq2::copy(bcsize+6) } // alpha
-            { Fq2::copy(bcsize+6) } // bias
-            { Fq2::copy(4 + 7) } // p_dash
-            { hinted_ell_tangent }
-            { Fq2::toaltstack() } // le.0
-            { Fq2::toaltstack() } // le.1
-
-
-            { Fq2::copy(bcsize+4)} // bias
-            { Fq2::copy(bcsize+8)} // alpha
-            { Fq2::copy(bcsize+6)} // t.x
-            { hinted_double_line }
-            { Fq2::toaltstack() }
-            { Fq2::toaltstack()}
-            // { fq2_push_not_montgomery(tt.y) }
-            // { Fq2::equalverify() }
-            // { fq2_push_not_montgomery(tt.x) }
-            // { Fq2::equalverify() }
-
-            { Fq2::roll(bcsize+6) } // alpha tangent drop
-            { Fq2::drop() }
-            { Fq2::roll(bcsize+4) } // bias tangent drop
-            { Fq2::drop() }
-
-            // hinted check chord // t.x, t.y
-            { Fq2::copy(bcsize+6)} // alpha
-            { Fq2::copy(bcsize+6)} // bias
-            { Fq2::copy(8+1) } // q.x
-            { Fq2::copy(8+1) } // q.y
-            { hinted_check_chord_q }
-            { Fq2::copy(bcsize+6)} // alpha
-            { Fq2::copy(bcsize+6)} // bias
-            { Fq2::fromaltstack() }
-            { Fq2::fromaltstack() }
-            { Fq2::copy(2)} // t.x
-            { Fq2::toaltstack() }
-            { hinted_check_chord_t }
-
-
-            { Fq2::copy(bcsize+6) } // alpha
-            { Fq2::copy(bcsize+6) } // bias
-            { Fq2::copy(10+1) } // p_dash
-            { hinted_ell_chord }
-
-            { Fq2::roll(4+bcsize+4) } // bias
-            { Fq2::roll(6+bcsize+4) } // alpha
-            { Fq2::copy(4+4+4+1) } //q.x
-            { Fq2::fromaltstack() }
-            { hinted_add_line }
-
-            { Fq2::toaltstack() }//R
-            { Fq2::toaltstack() }
-            
-            { Fq2::toaltstack() } //le_add
-            { Fq2::toaltstack() } 
-
-  
-            { Fq::toaltstack() } //hashes
-            { Fq::toaltstack() }
-            { Fq::toaltstack() }
-            { Fq2::drop() }
-            { Fq2::drop() }
-            { Fq2::drop() }
-
-            //T
-            {Fq::roll(1)}
-            {Fq::roll(2)}
-            {Fq::roll(3)}
-            { Fq::toaltstack() }
-            { Fq::toaltstack() }
-            { Fq::toaltstack() }
-            {unpack_u32_to_u8()} // 0
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            {hash_128b_168k.clone()}
-
-            // fetch 1 hash
-            { Fq::fromaltstack()} // aux_hash
-            {unpack_u32_to_u8()}
-            {hash_64b_75k.clone()}
-            {pack_u8_to_u32()}
-            { Fq::fromaltstack()} //input_hash
-            {Fq2::drop()} //{Fq::equalverify(1, 0)}
-
-            for i in 0..13 {
-                {Fq::fromaltstack()}
-            }
-
-            // Hash le
-            {Fq::roll(1)}
-            {Fq::roll(2)}
-            {Fq::roll(3)}
-            { Fq::toaltstack() }
-            { Fq::toaltstack() }
-            { Fq::toaltstack() }
-            {unpack_u32_to_u8()} // 0
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            {hash_128b_168k.clone()}
-            {pack_u8_to_u32()}
-            {Fq::toaltstack()}
-            
-            {Fq::roll(1)}
-            {Fq::roll(2)}
-            {Fq::roll(3)}
-            { Fq::toaltstack() }
-            { Fq::toaltstack() }
-            { Fq::toaltstack() }
-            {unpack_u32_to_u8()} // 0
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            {hash_128b_168k.clone()}
-            {pack_u8_to_u32()}
-            {Fq::toaltstack()}
-
-            {Fq::roll(1)}
-            {Fq::roll(2)}
-            {Fq::roll(3)}
-            { Fq::toaltstack() }
-            { Fq::toaltstack() }
-            { Fq::toaltstack() }
-            {unpack_u32_to_u8()} // 0
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            { Fq::fromaltstack()}
-            {unpack_u32_to_u8()}
-            {hash_128b_168k.clone()}
-            //{nibbles_to_fq()}
-
-            // bring back hash
-            { Fq::fromaltstack()}
-            { unpack_u32_to_u8()}
-            {hash_64b_75k.clone()}
-            { Fq::fromaltstack()}
-            { unpack_u32_to_u8()}
-            {hash_64b_75k.clone()}
-            {pack_u8_to_u32()}
-
-            {Fq2::drop()} //{Fq::equalverify(1, 0)}
-            OP_TRUE
-        };
-
-        let len = script.len();
-        let res = execute_script(script);
-        assert!(res.success);
-        for i in 0..res.final_stack.len() {
-            println!("{i:} {:?}", res.final_stack.get(i));
-        }
-        println!("script {} stack {}", len, res.stats.max_nb_stack_items);
     }
 
     #[test]
